Change CALPHAD grid dtype from float32 to float64; remove dead astype call
float32 grid truncates stored chemical potentials and mobilities to ~7
significant digits before the interpolation arithmetic. float64 keeps
full precision end-to-end. Memory cost doubles (~380 MB → ~760 MB at
resolution=0.01) but correctness is guaranteed.

The .astype(field_dtype, copy=False) in the main loop was a no-op
(interpolator_nb already returns float64; copy=False returns the same
object when dtype matches) — removed.

+++ examples/simulate.py
@@ -227,7 +227,7 @@ def main():
             current_composition, data_grid,
             resolution=cfg.resolution,
             tree=tree, calphad_values=calphad_values,
-        ).astype(field_dtype, copy=False)
+        )
 
         chemical_potentials = interpolated_data[..., :n_comp]    # J/mol
         mobilities          = interpolated_data[..., n_comp:]    # m² mol / (J s)
